.claude/skills/databricks-notebooks/scripts/requirements.txt
databricks-sdk>=0.20.0
python-dotenv>=1.0.0
